refactor: MessageSummary 컴포넌트 공용화로 재구현
