--- src/components/ListPage/MessageSummary/MessageSummary.module.css
@@ -1,38 +1,0 @@
-.recentMessages {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  row-gap: 1rem;
-}
-
-.profile-image-wrapper {
-  display: block;
-}
-
-.profile-image {
-  position: relative;
-  border: 1px solid var(--Gray-20);
-  border-radius: 50%;
-  width: 2.8rem;
-  height: 2.8rem;
-}
-
-.profile-image.second {
-  z-index: 2;
-  left: -1.3rem;
-}
-
-.profile-image.third {
-  z-index: 3;
-  left: -2.6rem;
-}
-
-.profile-image-last {
-  display: inline-block;
-  background-color: var(--surface);
-  font-size: 1.2rem;
-  font-weight: 500;
-  padding: 0.6rem;
-  z-index: 4;
-  left: -2.6rem;
-}
