--- src/components/ListPage/RecentCount/RecentCount.module.css
@@ -1,13 +1,0 @@
-.recentCount {
-  display: block;
-}
-
-.text {
-  font-size: 1.3rem;
-  font-weight: 400;
-}
-
-.number {
-  font-size: 1.3rem;
-  font-weight: 700;
-}
