+++ src/components/MessageSummary/MessageSummary.module.css
@@ -1,10 +1,16 @@
 .RecentMessages {
   display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  row-gap: 1rem;
+}
+
+.RecentMessages.PostPage {
   flex-direction: row;
   align-items: center;
 }
 
-.ProfileImageWrapper {
+.RrofileImageWrapper {
   display: block;
 }
 
@@ -26,17 +32,19 @@
   left: -2.6rem;
 }
 
-.ProfileImage.Last {
+.ProfileImageLast {
   display: inline-block;
-  background-color: var(--surface);
+  position: relative;
+  background-color: white;
   font-size: 1.2rem;
   font-weight: 500;
   padding: 0.6rem;
   z-index: 4;
-  left: -2.6rem;
+  left: -4rem;
 }
-@media (max-width: 1248px) {
-  .RecentMessages {
+
+@media (max-width: 1199px) {
+  .RecentMessages.PostPage {
     display: none;
   }
 }
