+++ src/components/RecentCount/RecentCount.js
@@ -2,9 +2,9 @@ import styles from "./RecentCount.module.css";
 
 function RecentCount({ recentCount }) {
   return (
-    <div className={styles.RecentCount}>
-      <span className={styles.Number}>{recentCount}</span>
-      <span className={styles.Text}>명이 작성했어요!</span>
+    <div className={styles.recentCount}>
+      <span className={styles.number}>{recentCount}</span>
+      <span className={styles.text}>명이 작성했어요!</span>
     </div>
   );
 }
